adding authentication to profiles router, moving github profiles to internal
