+++ internal/github/user.go
@@ -2,105 +2,115 @@ package githubAPI
 
 import (
 	"context"
+	"errors"
+	"net/http"
 
+	"github.com/GitCollabCode/GitCollab/internal/gitauth"
 	"github.com/google/go-github/github"
-	"golang.org/x/oauth2"
 )
 
-type GitHubUserAPI struct {
-	client *github.Client
-}
-
-func NewGitHubUserAPI(gitHubToken oauth2.TokenSource) *GitHubUserAPI {
-	tc := oauth2.NewClient(context.Background(), gitHubToken)
-	return &GitHubUserAPI{github.NewClient(tc)}
+func getGitClientFromContext(r *http.Request) *github.Client {
+	client := r.Context().Value(gitauth.ContextGitClient)
+	if client == nil { // might be able to remove?
+		return nil
+	}
+	// can remove above check if this cast returns nil, and does not
+	// cause mem fault
+	return client.(*github.Client)
 }
 
 /*
  * Get User ID from users github
  */
-func (g *GitHubUserAPI) GetUserID(client *github.Client) (int, error) {
-	user, _, err := client.Users.Get(context.Background(), "")
-	if err != nil {
-		return 0, err
+func GetUserID(r *http.Request) (int, error) {
+	client := getGitClientFromContext(r)
+	if client == nil {
+		return -1, errors.New("could not get client from context")
 	}
+	user, _, err := client.Users.Get(context.Background(), "")
 	return int(*user.ID), err
 }
 
 /*
  * Get Github Username from users GitHub
  */
-func (g *GitHubUserAPI) GetGitUserName(client *github.Client) (string, error) {
-
-	user, _, err := client.Users.Get(context.Background(), "")
-	if err != nil {
-		return "", err
+func GetGitUserName(r *http.Request) (string, error) {
+	client := getGitClientFromContext(r)
+	if client == nil {
+		return "", errors.New("could not get client from context")
 	}
+	user, _, err := client.Users.Get(context.Background(), "")
 	return *user.Name, err
 }
 
 /*
  * Get Icon URL from users github
  */
-func (g *GitHubUserAPI) GetUserIconURL(client *github.Client) (string, error) {
-	user, _, err := client.Users.Get(context.Background(), "")
-	if err != nil {
-		return "", err
+func GetUserIconURL(r *http.Request) (string, error) {
+	client := getGitClientFromContext(r)
+	if client == nil {
+		return "", errors.New("could not get client from context")
 	}
+	user, _, err := client.Users.Get(context.Background(), "")
 	return *user.AvatarURL, err
 }
 
 /*
  * Get Bio from users github
  */
-func (g *GitHubUserAPI) GetUserBio(client *github.Client) (string, error) {
-	user, _, err := client.Users.Get(context.Background(), "")
-	if err != nil {
-		return "", err
+func GetUserBio(r *http.Request) (string, error) {
+	client := getGitClientFromContext(r)
+	if client == nil {
+		return "", errors.New("could not get client from context")
 	}
+	user, _, err := client.Users.Get(context.Background(), "")
 	return *user.Bio, err
 }
 
 /*
  * Get Account name, not user name from users github
  */
-func (g *GitHubUserAPI) GetAccountName(client *github.Client) (string, error) {
-	user, _, err := client.Users.Get(context.Background(), "")
-	if err != nil {
-		return "", err
+func GetAccountName(r *http.Request) (string, error) {
+	client := getGitClientFromContext(r)
+	if client == nil {
+		return "", errors.New("could not get client from context")
 	}
+	user, _, err := client.Users.Get(context.Background(), "")
 	return *user.Login, err
 }
 
 /*
  * Get number of followers from users github
  */
-func (g *GitHubUserAPI) GetNumFollowers(client *github.Client) (int, error) {
-	user, _, err := client.Users.Get(context.Background(), "")
-	if err != nil {
-		return 0, err
+func GetNumFollowers(r *http.Request) (int, error) {
+	client := getGitClientFromContext(r)
+	if client == nil {
+		return -1, errors.New("could not get client from context")
 	}
+	user, _, err := client.Users.Get(context.Background(), "")
 	return *user.Followers, err
 }
 
 /*
  * Get number of accounts being followed from users github
  */
-func (g *GitHubUserAPI) GetNumFollowing(client *github.Client) (int, error) {
-	user, _, err := client.Users.Get(context.Background(), "")
-	if err != nil {
-		return 0, err
+func GetNumFollowing(r *http.Request) (int, error) {
+	client := getGitClientFromContext(r)
+	if client == nil {
+		return -1, errors.New("could not get client from context")
 	}
+	user, _, err := client.Users.Get(context.Background(), "")
 	return *user.Following, err
 }
 
 /*
  * Get Company from users github
  */
-func (g *GitHubUserAPI) GetCompany(client *github.Client) (string, error) {
-	user, _, err := client.Users.Get(context.Background(), "")
-	if err != nil {
-		return "", err
+func GetCompany(r *http.Request) (string, error) {
+	client := getGitClientFromContext(r)
+	if client == nil {
+		return "", errors.New("could not get client from context")
 	}
+	user, _, err := client.Users.Get(context.Background(), "")
 	return *user.Company, err
 }
